requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-ulid==0.0.4
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.1
PyJWT==2.8.0
sqlparse==0.4.4
ulid-py==1.1.0
whitenoise==6.7.0
